# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CoursePulse AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CourseContext {
  id: number;
  week: number;
  topic: string;
  learningObjective: string;
  /** @nullable */
  assignment: string | null;
  /** @nullable */
  reading: string | null;
}

export interface CourseContextInput {
  week: number;
  topic: string;
  learningObjective: string;
  assignment?: string;
  reading?: string;
}

export interface CourseContextUpdate {
  week?: number;
  topic?: string;
  learningObjective?: string;
  assignment?: string;
  reading?: string;
}

export interface Reflection {
  id: number;
  week: number;
  topic: string;
  learningObjective: string;
  confidenceScore: number;
  reflectionText: string;
  supportRequested: boolean;
  createdAt: string;
}

export interface ReflectionInput {
  week: number;
  topic: string;
  learningObjective: string;
  confidenceScore: number;
  reflectionText: string;
  supportRequested: boolean;
  consentGiven: boolean;
}

export interface ClassifiedSignal {
  id: number;
  reflectionId: number;
  primarySignal: string;
  /** @nullable */
  secondarySignal: string | null;
  severityScore: number;
  themeSummary: string;
  recommendedAction: string;
}

export interface ReflectionWithSignal {
  reflection: Reflection;
  signal: ClassifiedSignal;
}

export interface SignalWithContext {
  id: number;
  reflectionId: number;
  primarySignal: string;
  /** @nullable */
  secondarySignal: string | null;
  severityScore: number;
  themeSummary: string;
  recommendedAction: string;
  week: number;
  topic: string;
  learningObjective: string;
}

export interface FacultyAction {
  id: number;
  week: number;
  topic: string;
  actionTaken: string;
  createdAt: string;
}

export interface FacultyActionInput {
  week: number;
  topic: string;
  actionTaken: string;
}

export interface FacultyVerifyInput {
  accessCode: string;
}

export interface FacultyVerifyResponse {
  success: boolean;
}

export interface SignalCount {
  signal: string;
  count: number;
}

export interface TopicConfusionItem {
  topic: string;
  avgSeverity: number;
  count: number;
  mostCommonSignal: string;
}

export interface ConfidenceTrendItem {
  week: number;
  avgConfidence: number;
  reflectionCount: number;
}

export interface DashboardStats {
  totalReflections: number;
  avgConfidenceScore: number;
  mostCommonSignal: string;
  /** @nullable */
  highestConcernTopic: string | null;
  percentRequestingSupport: number;
  signalDistribution: SignalCount[];
  topicConfusion: TopicConfusionItem[];
  confidenceTrend: ConfidenceTrendItem[];
  topConfusionThemes: string[];
  recommendedAdjustments: string[];
}