# lib/api-zod/src/generated/types/courseContext.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CoursePulse AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CourseContext {
  id: number;
  week: number;
  topic: string;
  learningObjective: string;
  /** @nullable */
  assignment: string | null;
  /** @nullable */
  reading: string | null;
}