# lib/api-zod/src/generated/types/courseContextInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CoursePulse AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CourseContextInput {
  week: number;
  topic: string;
  learningObjective: string;
  assignment?: string;
  reading?: string;
}