# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CoursePulse AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConfidenceTrendItem } from './confidenceTrendItem';
import type { SignalCount } from './signalCount';
import type { TopicConfusionItem } from './topicConfusionItem';

export interface DashboardStats {
  totalReflections: number;
  avgConfidenceScore: number;
  mostCommonSignal: string;
  /** @nullable */
  highestConcernTopic: string | null;
  percentRequestingSupport: number;
  signalDistribution: SignalCount[];
  topicConfusion: TopicConfusionItem[];
  confidenceTrend: ConfidenceTrendItem[];
  topConfusionThemes: string[];
  recommendedAdjustments: string[];
}