# lib/api-zod/src/generated/types/facultyActionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CoursePulse AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FacultyActionInput {
  week: number;
  topic: string;
  actionTaken: string;
}