# lib/api-zod/src/generated/types/reflection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CoursePulse AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Reflection {
  id: number;
  week: number;
  topic: string;
  learningObjective: string;
  confidenceScore: number;
  reflectionText: string;
  supportRequested: boolean;
  createdAt: string;
}