# lib/api-zod/src/generated/types/reflectionWithSignal.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CoursePulse AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClassifiedSignal } from './classifiedSignal';
import type { Reflection } from './reflection';

export interface ReflectionWithSignal {
  reflection: Reflection;
  signal: ClassifiedSignal;
}