# lib/api-zod/src/generated/types/signalWithContext.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CoursePulse AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SignalWithContext {
  id: number;
  reflectionId: number;
  primarySignal: string;
  /** @nullable */
  secondarySignal: string | null;
  severityScore: number;
  themeSummary: string;
  recommendedAction: string;
  week: number;
  topic: string;
  learningObjective: string;
}