# lib/api-zod/src/generated/types/topicConfusionItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CoursePulse AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TopicConfusionItem {
  topic: string;
  avgSeverity: number;
  count: number;
  mostCommonSignal: string;
}